linkedin_jobs_notification/pyproject.toml
[project]
name = "linkedin-jobs-notification"
version = "0.1.0"
description = "Cron-based LinkedIn job watcher with Telegram notifications"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "linkedin-scraper",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.22.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = []

[tool.uv.sources]
linkedin-scraper = { path = "../linkedin_scraper", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "respx>=0.22.0",
]
